Fix messages not going through
When there are many borked users in a room, the client will request all their public keys
before sending a message. This clashes with server-side rate limiting, and results in the
actual message being dropped. The functionality to auto-request keys was added in 2.4.0
and is no longer needed anyway due to protocol changes.